LayoutTests/imported/w3c/web-platform-tests/html/editing/editing-0/autocapitalization/autocapitalize-expected.txt



FAIL Test that the autocapitalize is avaible on HTMLInputElement. assert_true: expected true got false
FAIL Test that the autocapitalize is avaible on HTMLTextAreaElement. assert_true: expected true got false
FAIL Test that the autocapitalize is avaible on div. assert_true: expected true got false
FAIL Test deprecated values of autocapitalize. assert_equals: expected "sentences" but got "on"
FAIL Test reflection of autocapitalize. assert_equals: expected (string) "" but got (undefined) undefined
FAIL Test that the IDL attribute returns the empty string if the content attribute is not set. assert_equals: expected (string) "" but got (undefined) undefined
FAIL Test inheriting values from a form. assert_equals: button element with form parent with "null" and form element with "null" expected (string) "" but got (undefined) undefined
FAIL Verify that even input types that are never autocapitalized support the IDL interface. assert_equals: expected (string) "sentences" but got (undefined) undefined
